Moved Result into its own component
