+++ src/components/Result/index.js
@@ -1,0 +1,3 @@
+import Result from './Result';
+
+export default Result;
